Alineacion de motores

+++ src/control.h
@@ -84,8 +84,8 @@ extern "C" {
     }STA_data;
     
     /**********Control Specific Functions**********/
-    void Control_initialize(uint16_t reference,STA_data *SMC_ST_data ,as5600_sensor *sensor, uint8_t motor_number, float const_c1, float const_c2, float const_b);
-    void Control_initialize_As5048(uint16_t reference, STA_data *SMC_ST_data ,as5048a_sensor *sensor, uint8_t motor_number, float const_c1, float const_c2, float const_b);
+    void Control_initialize(float reference,STA_data *SMC_ST_data ,as5600_sensor *sensor, uint8_t motor_number, float const_c1, float const_c2, float const_b);
+    void Control_initialize_As5048(float reference, STA_data *SMC_ST_data ,as5048a_sensor *sensor, uint8_t motor_number, float const_c1, float const_c2, float const_b);
     //void Control_PID(float kp, float ki, float kd);
     void Control_UpdateSpeedAcceleration(void);
     //void Control_UpdateDirection(float pwm_duty);
